Updated runtime to .Net 8.
