+++ SamplePeteWebApp/Properties/launchSettings.json
@@ -17,7 +17,7 @@
     },
     "SamplePeteWebApp": {
       "commandName": "Project",
-      "dotnetRunMessages": "true",
+      "dotnetRunMessages": true,
       "launchBrowser": true,
       "applicationUrl": "https://localhost:5001;http://localhost:5000",
       "environmentVariables": {
