+++ SamplePeteWebApp/Views/Shared/_Layout.cshtml
@@ -50,7 +50,7 @@
 
     <footer class="border-top footer text-muted footer-bg-color">
         <div class="container footer-bg-color">
-            &copy; 2021 - SamplePeteWebApp
+            &copy; @DateTime.Now.Year - SamplePeteWebApp
         </div>
     </footer>
 </body>
